Add files via upload
The relevant dataset represents the demographic and local financial information of existing customers in a specific region of a bank.The data set consists of 23,000 observations from 35 different variables.

The problem is although there are no credit card recommendation systems that the bank currently uses, we know that the basic rules sets are used in new applications. In fact, the bank should be able to anticipate and identify its potential customers based on the purpose of increasing the risk appetite. For this reason, this study is mainly based on the problem of predetermining the prospective customers who are newcomers to the credit card product.

The objective of this study is to estimate the appropriate credit card users from existing customers by using the data of credit card customers belonging to a specific time period.

Basically, two types of information were collected;
1. The first is the demographic information of the customers. (Ex., Customer Total Revenue, Customer Professionals, Customer Age etc.)
2. Second, the number and details of the products used by customers in the bank. (Ex., Total Products Used, Credit Card Use, Credit Usage etc.)

The description and coding of each variable is described in more detail in the following sections.